+++ moirai-core/src/lib.rs
@@ -7,6 +7,7 @@
 //! and scheduling interfaces.
 
 #![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(nightly_tls_active, feature(thread_local))]
 #![warn(missing_docs)]
 #![warn(clippy::all)]
 #![warn(clippy::pedantic)]